split most types out of solenoid configs, so it's more readable
